remove unnecessary conditionals

+++ src/core/hw/gpu/engine_3d/geometry_engine.cpp
@@ -325,10 +325,12 @@ void GeometryEngine::DoSwapBuffers() {
         if (vertex_ram[i].w != 0) {
             u16 screen_width = (gpu->render_engine.screen_x2 - gpu->render_engine.screen_x1 + 1) & 0x1FF;
             u16 screen_height = (gpu->render_engine.screen_y2 - gpu->render_engine.screen_y1 + 1) & 0xFF;
-            u16 render_x = ((s64)vertex_ram[i].x + vertex_ram[i].w) * screen_width / (2 * vertex_ram[i].w) + gpu->render_engine.screen_x1;
-            u16 render_y = (-(s64)vertex_ram[i].y + vertex_ram[i].w) * screen_height / (2 * vertex_ram[i].w) + gpu->render_engine.screen_y1;
+            s64 render_x = ((s64)vertex_ram[i].x + vertex_ram[i].w) * screen_width / (2 * vertex_ram[i].w) + gpu->render_engine.screen_x1;
+            s64 render_y = (-(s64)vertex_ram[i].y + vertex_ram[i].w) * screen_height / (2 * vertex_ram[i].w) + gpu->render_engine.screen_y1;
             render_x &= 0x1FF;
             render_y &= 0xFF;
+
+            // printf("%d %d\n", gpu->render_engine.screen_x1, gpu->render_engine.screen_y2);
 
             Vertex render_vertex;
